FileMisc.forceDelete() is now okay with files which don't exist.

+++ src/main/java/com/diffplug/gradle/FileMisc.java
@@ -73,10 +73,12 @@ public class FileMisc {
 		});
 	}
 
-	/** Calls {@link FileUtils#forceDelete(File)} and throws an exception if it fails. */
+	/** Calls {@link FileUtils#forceDelete(File)} and throws an exception if it fails.  If the file doesn't exist at all, that's fine. */
 	public static void forceDelete(File f) {
 		retry(f, file -> {
-			FileUtils.forceDelete(f);
+			if (file.exists()) {
+				FileUtils.forceDelete(f);
+			}
 			return null;
 		});
 	}
